refactored code into component
